third_party/WebKit/LayoutTests/http/tests/worklet/webexposed/global-interface-listing-paint-worklet-expected.txt

CONSOLE MESSAGE: line 137: interface ByteLengthQueuingStrategy
CONSOLE MESSAGE: line 137:     method constructor
CONSOLE MESSAGE: line 137:     method size
CONSOLE MESSAGE: line 137: interface CountQueuingStrategy
CONSOLE MESSAGE: line 137:     method constructor
CONSOLE MESSAGE: line 137:     method size
CONSOLE MESSAGE: line 137: interface DataView
CONSOLE MESSAGE: line 137:     getter buffer
CONSOLE MESSAGE: line 137:     getter byteLength
CONSOLE MESSAGE: line 137:     getter byteOffset
CONSOLE MESSAGE: line 137:     method constructor
CONSOLE MESSAGE: line 137:     method getFloat32
CONSOLE MESSAGE: line 137:     method getFloat64
CONSOLE MESSAGE: line 137:     method getInt16
CONSOLE MESSAGE: line 137:     method getInt32
CONSOLE MESSAGE: line 137:     method getInt8
CONSOLE MESSAGE: line 137:     method getUint16
CONSOLE MESSAGE: line 137:     method getUint32
CONSOLE MESSAGE: line 137:     method getUint8
CONSOLE MESSAGE: line 137:     method setFloat32
CONSOLE MESSAGE: line 137:     method setFloat64
CONSOLE MESSAGE: line 137:     method setInt16
CONSOLE MESSAGE: line 137:     method setInt32
CONSOLE MESSAGE: line 137:     method setInt8
CONSOLE MESSAGE: line 137:     method setUint16
CONSOLE MESSAGE: line 137:     method setUint32
CONSOLE MESSAGE: line 137:     method setUint8
CONSOLE MESSAGE: line 137: interface PaintRenderingContext2D
CONSOLE MESSAGE: line 137:     getter currentTransform
CONSOLE MESSAGE: line 137:     getter fillStyle
CONSOLE MESSAGE: line 137:     getter globalAlpha
CONSOLE MESSAGE: line 137:     getter globalCompositeOperation
CONSOLE MESSAGE: line 137:     getter imageSmoothingEnabled
CONSOLE MESSAGE: line 137:     getter imageSmoothingQuality
CONSOLE MESSAGE: line 137:     getter lineCap
CONSOLE MESSAGE: line 137:     getter lineDashOffset
CONSOLE MESSAGE: line 137:     getter lineJoin
CONSOLE MESSAGE: line 137:     getter lineWidth
CONSOLE MESSAGE: line 137:     getter miterLimit
CONSOLE MESSAGE: line 137:     getter shadowBlur
CONSOLE MESSAGE: line 137:     getter shadowColor
CONSOLE MESSAGE: line 137:     getter shadowOffsetX
CONSOLE MESSAGE: line 137:     getter shadowOffsetY
CONSOLE MESSAGE: line 137:     getter strokeStyle
CONSOLE MESSAGE: line 137:     method arc
CONSOLE MESSAGE: line 137:     method arcTo
CONSOLE MESSAGE: line 137:     method beginPath
CONSOLE MESSAGE: line 137:     method bezierCurveTo
CONSOLE MESSAGE: line 137:     method clearRect
CONSOLE MESSAGE: line 137:     method clip
CONSOLE MESSAGE: line 137:     method closePath
CONSOLE MESSAGE: line 137:     method constructor
CONSOLE MESSAGE: line 137:     method createLinearGradient
CONSOLE MESSAGE: line 137:     method createPattern
CONSOLE MESSAGE: line 137:     method createRadialGradient
CONSOLE MESSAGE: line 137:     method drawImage
CONSOLE MESSAGE: line 137:     method ellipse
CONSOLE MESSAGE: line 137:     method fill
CONSOLE MESSAGE: line 137:     method fillRect
CONSOLE MESSAGE: line 137:     method getLineDash
CONSOLE MESSAGE: line 137:     method isPointInPath
CONSOLE MESSAGE: line 137:     method isPointInStroke
CONSOLE MESSAGE: line 137:     method lineTo
CONSOLE MESSAGE: line 137:     method moveTo
CONSOLE MESSAGE: line 137:     method quadraticCurveTo
CONSOLE MESSAGE: line 137:     method rect
CONSOLE MESSAGE: line 137:     method resetTransform
CONSOLE MESSAGE: line 137:     method restore
CONSOLE MESSAGE: line 137:     method rotate
CONSOLE MESSAGE: line 137:     method save
CONSOLE MESSAGE: line 137:     method scale
CONSOLE MESSAGE: line 137:     method setLineDash
CONSOLE MESSAGE: line 137:     method setTransform
CONSOLE MESSAGE: line 137:     method stroke
CONSOLE MESSAGE: line 137:     method strokeRect
CONSOLE MESSAGE: line 137:     method transform
CONSOLE MESSAGE: line 137:     method translate
CONSOLE MESSAGE: line 137:     setter currentTransform
CONSOLE MESSAGE: line 137:     setter fillStyle
CONSOLE MESSAGE: line 137:     setter globalAlpha
CONSOLE MESSAGE: line 137:     setter globalCompositeOperation
CONSOLE MESSAGE: line 137:     setter imageSmoothingEnabled
CONSOLE MESSAGE: line 137:     setter imageSmoothingQuality
CONSOLE MESSAGE: line 137:     setter lineCap
CONSOLE MESSAGE: line 137:     setter lineDashOffset
CONSOLE MESSAGE: line 137:     setter lineJoin
CONSOLE MESSAGE: line 137:     setter lineWidth
CONSOLE MESSAGE: line 137:     setter miterLimit
CONSOLE MESSAGE: line 137:     setter shadowBlur
CONSOLE MESSAGE: line 137:     setter shadowColor
CONSOLE MESSAGE: line 137:     setter shadowOffsetX
CONSOLE MESSAGE: line 137:     setter shadowOffsetY
CONSOLE MESSAGE: line 137:     setter strokeStyle
CONSOLE MESSAGE: line 137: interface PaintWorkletGlobalScope : WorkletGlobalScope
CONSOLE MESSAGE: line 137:     method constructor
CONSOLE MESSAGE: line 137: interface ReadableStream
CONSOLE MESSAGE: line 137:     getter locked
CONSOLE MESSAGE: line 137:     method cancel
CONSOLE MESSAGE: line 137:     method constructor
CONSOLE MESSAGE: line 137:     method getReader
CONSOLE MESSAGE: line 137:     method tee
CONSOLE MESSAGE: line 137: interface WorkletGlobalScope
CONSOLE MESSAGE: line 137:     getter console
CONSOLE MESSAGE: line 137:     method constructor
CONSOLE MESSAGE: line 137:     setter console
CONSOLE MESSAGE: line 137: global object
CONSOLE MESSAGE: line 137:     method gc
CONSOLE MESSAGE: line 137:     method registerPaint
